Update azure-pipeline.yml
Bump node js version

Fix prep command on Linux
