+++ azure-pipelines-ci.yml
@@ -1,0 +1,13 @@
+steps:
+- task: NodeTool@0
+  inputs:
+    versionSpec: '8.12.0'
+  displayName: 'Install Node.js'
+
+- script: |
+    npm ci
+  displayName: 'Install dependencies'
+
+- script: |
+    npm run build:prod
+  displayName: 'Compile and package'
