+++ extension.json
@@ -2,7 +2,7 @@
     "manifestVersion": 1,
     "id": "samples",
     "publisher": "ms-samples",
-    "version": "1.0.145",
+    "version": "1.0.156",
     "name": "Extension Sample",
     "description": "Azure DevOps extension sample",
     "categories": [
@@ -18,7 +18,7 @@
     },
     "content": {
         "details": {
-            "path": "overview.md"
+            "path": "README.md"
         }
     },
     "files": [
