--- overview.md
@@ -1,1 +1,0 @@
-Sample Azure DevOps extension
